examples/code-agent/requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0

# E2B Sandbox
e2b>=1.0.0

# API Server
fastapi>=0.115.0
uvicorn>=0.32.0

# Web UI
streamlit>=1.40.0

# Utilities
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Async support
aiofiles>=24.1.0

# Vector search / Indexing
faiss-cpu>=1.8.0
openai>=1.50.0
numpy>=1.26.0
